Removed all_types, updated basic_types and extend_types.
