+++ pytdml/io/tdml_writers.py
@@ -31,6 +31,8 @@
 # ------------------------------------------------------------------------------
 import json
 from typing import Union
+
+from pytdml.io import read_from_json
 from pytdml.type import TrainingDataset, EOTrainingDataset
