+++ pytdml/type/extended_types.py
@@ -31,99 +31,177 @@
 # ------------------------------------------------------------------------------
 
 import json
-
+import copy
+import geojson
 from geojson import Feature
 from typing import List, Union, Optional, Literal
 from pydantic import Field, field_validator
 
-from pytdml.type._utils import _validate_image_format, _validate_date
-from pytdml.type.basic_types import Label, TrainingDataset, TrainingData, Task, MD_Band, Extent, BoundingBox
+from pytdml.type._utils import _validate_date, _validate_image_format, to_interior_class, list_to_interior_class
+from pytdml.type.basic_types import AI_Label, TrainingDataset, AI_TrainingData, MD_Band, EX_Extent, CI_Citation, AI_Labeling, DataQuality, AI_Task, NamedValue, MD_Scope, AI_MetricsInLiterature, AI_TDChangeset
 
 
-class PixelLabel(Label):
+class AI_PixelLabel(AI_Label):
     """
     Extended label type for pixel level training data
     """
+
     type: Literal["AI_PixelLabel"]
-    image_URL: List[str] = Field(min_items=1)
-    image_format: List[str] = Field(min_items=1)
+    image_URL: List[str] = Field(min_length=1)
+    image_format: List[str] = Field(min_length=1)
 
     @field_validator("image_format")
     def validate_image_format(cls, v):
         valid_format = []
         for item in v:
-            valid_format.append(_validate_image_format(item))
+            if _validate_image_format(item):
+                valid_format.append(item)
         return valid_format
 
+    def to_dict(self):
+        return self.model_dump(by_alias=True, exclude_none=True)
 
-class ObjectLabel(Label):
+    @staticmethod
+    def from_dict(json_dict):
+        new_dict = copy.deepcopy(json_dict)
+        return AI_PixelLabel(**new_dict)
+
+
+class AI_ObjectLabel(AI_Label):
     """
     Extended label type for object level training data
     """
+
     type: Literal["AI_ObjectLabel"]
     object: Feature
     label_class: str = Field(alias="class")
 
-    date_time: Optional[str]
-    bbox_type: Optional[str]
+    date_time: Optional[str] = None
+    bbox_type: Optional[str] = None
 
     @field_validator("date_time")
     def validate_date_time(cls, v):
         return _validate_date(v)
 
+    def to_dict(self):
+        return self.model_dump(by_alias=True, exclude_none=True)
 
-class SceneLabel(Label):
+    @staticmethod
+    def from_dict(json_dict):
+        new_dict = copy.deepcopy(json_dict)
+        if new_dict.__contains__('object'):
+            new_dict["object"] = geojson.loads(json.dumps(json_dict["object"]))
+        else:
+            print("Parameter \"object\" must be provided.")
+            exit()
+        return AI_ObjectLabel(**new_dict)
+
+
+class AI_SceneLabel(AI_Label):
     """
     Extended label type for scene level training data
     """
+
     type: Literal["AI_SceneLabel"]
     label_class: str = Field(alias="class")
 
+    def to_dict(self):
+        return self.model_dump(by_alias=True, exclude_none=True)
 
-class EOTask(Task):
+    @staticmethod
+    def from_dict(json_dict):
+        new_dict = copy.deepcopy(json_dict)
+        return AI_SceneLabel(**new_dict)
+
+
+class AI_EOTask(AI_Task):
     """
     Extended task type for EO training data
     """
     type: Literal["AI_EOTask"]
-    taskType: str
+    task_type: str
+
+    def to_dict(self):
+        return self.model_dump(by_alias=True, exclude_none=True)
+
+    @staticmethod
+    def from_dict(json_dict):
+        new_dict = copy.deepcopy(json_dict)
+        return AI_EOTask(**new_dict)
 
 
-class EOTrainingData(TrainingData):
+class AI_EOTrainingData(AI_TrainingData):
     """
     Extended training data type for EO training data
     """
+
     type: Literal["AI_EOTrainingData"]
-    dataURL: List[str] = Field(min_items=1)  # That one should be uri-format
+    data_URL: List[str] = Field(min_length=1)  # That one should be uri-format
+    labels: List[Union[AI_Label, AI_PixelLabel, AI_ObjectLabel, AI_SceneLabel]]
 
-    extent: Optional[Union[Extent, BoundingBox]]
-    date_time: Optional[List[str]] = []
+    extent: Optional[Union[EX_Extent, List[Union[int, float]]]] = None
+    data_time: Optional[List[str]] = None
 
-    @field_validator("date_time")
+    @field_validator("data_time")
     def validate_data_time(cls, v):
-        validated_data = []
+        validated_date = []
         for item in v:
-            validated_data.append(_validate_date(item))
-        return validated_data
+            validated_date.append(_validate_date(item))
+        return validated_date
+
+    def to_dict(self):
+        return self.model_dump(by_alias=True, exclude_none=True)
+
+    @staticmethod
+    def from_dict(json_dict):
+        new_dict = copy.deepcopy(json_dict)
+        labels = new_dict['labels']
+        for i in range(len(labels)):
+            if labels[i]["type"] == "AI_AbstractLabel":
+                labels[i] = AI_Label.from_dict(labels[i])
+            elif labels[i]["type"] == "AI_PixelLabel":
+                labels[i] = AI_PixelLabel.from_dict(labels[i])
+            elif labels[i]["type"] == "AI_ObjectLabel":
+                labels[i] = AI_ObjectLabel.from_dict(labels[i])
+            else:
+                labels[i] = AI_SceneLabel.from_dict(labels[i])
+        new_dict['labels'] = labels
+        if new_dict.__contains__('dataSources'):
+            list_to_interior_class(new_dict, "dataSources", CI_Citation)
+        if new_dict.__contains__('labeling'):
+            list_to_interior_class(new_dict, "labeling", AI_Labeling)
+        if new_dict.__contains__('quality'):
+            list_to_interior_class(new_dict, "quality", DataQuality)
+
+        if new_dict.__contains__('extent'):
+            extent = new_dict['extent']
+            for i in range(len(extent)):
+                if EX_Extent.can_build_from_data(extent[i]):
+                    extent[i] = EX_Extent.from_dict(extent[i])
+                else:
+                    pass
+            new_dict['extent'] = extent
+        return AI_EOTrainingData(**new_dict)
 
 
 class EOTrainingDataset(TrainingDataset):
     """
     Extended training dataset type for EO training dataset
     """
+
     type: Literal["AI_EOTrainingDataset"]
+    tasks: List[AI_EOTask] = Field(min_length=1)
+    data: List[AI_EOTrainingData] = Field(min_length=1)
     # For Convinience, we allow the user to specify the bands by name
 
-    bands: Optional[List[Union[str, MD_Band]]] = []
-    extent: Optional[Extent]
-    imageSize: Optional[str] = ""
-    tasks: List[EOTask] = Field(min_items=1)
-    data: List[EOTrainingData] = Field(min_items=1)
+    bands: Optional[List[MD_Band]] = None
+    extent: Optional[EX_Extent] = None
+    image_size: Optional[str] = None
 
+    def to_dict(self):
+        return self.model_dump(by_alias=True, exclude_none=True)
 
-if __name__ == "__main__":
-    tdml_path = r"C:\Users\zhaoyan\Desktop\TrainingDML-AI-master-use-cases-examples\use-cases\examples\1.0\WHU-building.json"
-    with open(tdml_path, 'r') as f:
-        data = json.load(f)
-
-    td = EOTrainingDataset(**data).dict(by_alias=True,exclude_none=True)
-    print(td['data'])
+    @staticmethod
+    def from_dict(json_dict):
+        new_dict = copy.deepcopy(json_dict)
+        return EOTrainingDataset(**new_dict)
